ProiectCloud/requirements.txt
Django
gunicorn
twitter
social-auth-app-django
python-twitter
azure-storage-file
textblob
urlextract
tweepy
